修改输入，支持幂指数平滑采样
