+++ models/biaffine_model.py
@@ -6,7 +6,7 @@ import pathlib
 import torch
 import torch.nn as nn
 
-from utils.input_utils.bertology.bert_input_utils import load_bert_tokenizer, load_and_cache_examples, get_data_loader
+from utils.input_utils.bertology.input_utils import load_bert_tokenizer, load_and_cache_examples, get_data_loader
 from utils.input_utils.graph_vocab import GraphVocab
 from modules.bertology_encoder import BERTologyEncoder
 from modules.biaffine import DeepBiaffineScorer, DirectBiaffineScorer
@@ -66,52 +66,3 @@ class BiaffineDependencyModel(BaseModel):
         unlabeled_scores = self.unlabeled_biaffine(encoder_output, encoder_output).squeeze(3)
         labeled_scores = self.labeled_biaffine(encoder_output, encoder_output)
         return unlabeled_scores, labeled_scores
-
-
-if __name__ == '__main__':
-    class Args():
-        def __init__(self):
-            self.bert_path = '/home/liangs/disk/data/bertology-base-chinese'
-            self.data_dir = '../dataset'
-            self.train_file = 'test.conllu'
-            self.max_seq_length = 10
-            self.encoder_type = 'bertology'
-            self.root_representation = 'unused'
-            self.graph_vocab_file = '../dataset/graph_vocab.txt'
-            self.cuda = False
-            self.bert_chinese_word_select = 's+e'
-            self.bert_output_mode = 'last_four_sum'
-
-            # for Biaffine
-            self.biaffine_hidden_dim = 300
-            self.biaffine_dropout = 0.1
-
-            # for loss:
-            self.learned_loss_ratio = True,
-            self.label_loss_ratio = 0.5
-
-
-    args = Args()
-
-    if args.encoder_type == 'bertology':
-        args.encoder_output_dim = 768
-
-    tokenizer = load_bert_tokenizer('/home/liangs/disk/data/bertology-base-chinese', 'bertology')
-    vocab = GraphVocab('../dataset/graph_vocab.txt')
-    dataset, CoNLLU_file = load_and_cache_examples(args, vocab, tokenizer)
-    data_loader = get_data_loader(dataset, batch_size=2, evaluation=True)
-    # bertology = BERTTypeEncoder(no_cuda=True, bert_path=args.bert_path)
-    model = BiaffineDependencyModel(args)
-    print(model)
-    for batch in data_loader:
-        inputs = {
-            'input_ids': batch[0],
-            'attention_mask': batch[1],
-            'token_type_ids': batch[2] if args.encoder_type in ['bertology', 'xlnet'] else None,
-            'start_pos': batch[3],
-            'end_pos': batch[4],
-        }
-        # print(inputs)
-        # print(inputs['start_pos'])
-        output = model(inputs)
-        print(output)
